backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic>=2.7.4
python-dotenv==1.0.1
httpx==0.27.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
